requirements.txt
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
openai==0.27.8
requests==2.31.0
pyyaml==6.0
pydantic==2.0.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
